refactor(persistence-sqlite): decompose sqlite_store.rs into focused submodules (KAN-764)
