+++ crates/kanban-persistence-sqlite/src/sqlite_store/entities/column.rs
@@ -1,0 +1,35 @@
+use kanban_domain::{Column, KanbanResult};
+
+use crate::sqlite_store::helpers::{db_err, fmt_dt, required_str};
+use crate::sqlite_store::SqliteStore;
+
+impl SqliteStore {
+    pub(crate) async fn write_column_with_conn(
+        conn: &mut sqlx::SqliteConnection,
+        column: &Column,
+    ) -> KanbanResult<()> {
+        sqlx::query(
+            "INSERT INTO columns (id, board_id, name, position, wip_limit, created_at, updated_at)
+             VALUES (?, ?, ?, ?, ?, ?, ?)
+             ON CONFLICT(id) DO UPDATE SET
+                board_id=excluded.board_id, name=excluded.name,
+                position=excluded.position, wip_limit=excluded.wip_limit,
+                updated_at=excluded.updated_at",
+        )
+        .bind(column.id.to_string())
+        .bind(column.board_id.to_string())
+        .bind(required_str(&column.name, "column.name")?)
+        .bind(column.position)
+        .bind(column.wip_limit)
+        .bind(fmt_dt(&column.created_at))
+        .bind(fmt_dt(&column.updated_at))
+        .execute(&mut *conn)
+        .await
+        .map_err(db_err)?;
+        Ok(())
+    }
+
+    pub(crate) async fn write_column_async(&self, column: &Column) -> KanbanResult<()> {
+        Self::write_column_with_conn(&mut *self.pool.acquire().await.map_err(db_err)?, column).await
+    }
+}
