+++ crates/kanban-persistence-sqlite/src/sqlite_store/entities/mod.rs
@@ -1,0 +1,5 @@
+mod archived_card;
+mod board;
+mod card;
+mod column;
+mod sprint;
